* Skipping more than 2GB is possible (long long int instead of int)
git-svn-id: https://svn.gwserver.de:3691/hddwiper/trunk@36 c257ac6a-9548-4c2e-97fb-1c9d84f8ff2e

+++ HDDWiper.hpp
@@ -14,7 +14,7 @@ public:
 	static const unsigned int BLOCKSIZE=100*1024*1024;
 	static const unsigned int BUFFERSIZE=5; //in blocks
 
-	HDDWiper(const std::string &filename, const int skip=0);
+	HDDWiper(const std::string &filename, const long long int skip=0);
 
 	double getCurrentSpeed() const;
 
@@ -65,7 +65,7 @@ inline double HDDWiper::WipingThread::getCurrentSpeed() const
 	return _currentspeed;
 }
 
-inline HDDWiper::HDDWiper(const std::string &filename, const int skip)
+inline HDDWiper::HDDWiper(const std::string &filename, const long long int skip)
 	:_output(filename),_wipingthread(*this),_wipingthread_thread(boost::ref(_wipingthread))
 {
 	if(skip>0)
